Refactored as suggested by Kai
