+++ src/webgpu/api/validation/encoding/cmds/debug.spec.ts
@@ -12,7 +12,7 @@ Test Coverage:
 `;
 
 import { makeTestGroup } from '../../../../../common/framework/test_group.js';
-import { kEncoderTypes } from '../../util/command_buffer_maker.js';
+import { kEncoderTypes } from '../../../../util/command_buffer_maker.js';
 import { ValidationTest } from '../../validation_test.js';
 
 export const g = makeTestGroup(ValidationTest);
