+++ src/webgpu/api/validation/encoding/cmds/render/indirect_draw.spec.ts
@@ -4,7 +4,8 @@ Validation tests for drawIndirect/drawIndexedIndirect on render pass and render 
 
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
 import { GPUConst } from '../../../../../constants.js';
-import { kResourceStates, ValidationTest } from '../../../validation_test.js';
+import { kResourceStates } from '../../../../../gpu_test.js';
+import { ValidationTest } from '../../../validation_test.js';
 
 import { kRenderEncodeTypeParams } from './render.js';
