+++ src/webgpu/api/validation/encoding/cmds/render/render.ts
@@ -1,5 +1,5 @@
 import { kUnitCaseParamsBuilder } from '../../../../../../common/framework/params_builder.js';
-import { kRenderEncodeTypes } from '../../../util/command_buffer_maker.js';
+import { kRenderEncodeTypes } from '../../../../../util/command_buffer_maker.js';
 
 export const kRenderEncodeTypeParams = kUnitCaseParamsBuilder.combine(
   'encoderType',
