+++ src/webgpu/api/validation/encoding/cmds/render/setPipeline.spec.ts
@@ -3,7 +3,7 @@ Validation tests for setPipeline on render pass and render bundle.
 `;
 
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
-import { kRenderEncodeTypes } from '../../../util/command_buffer_maker.js';
+import { kRenderEncodeTypes } from '../../../../../util/command_buffer_maker.js';
 import { ValidationTest } from '../../../validation_test.js';
 
 import { kRenderEncodeTypeParams } from './render.js';
