+++ src/webgpu/api/validation/encoding/cmds/setBindGroup.spec.ts
@@ -13,11 +13,12 @@ TODO: merge these notes and implement.
 import { makeTestGroup } from '../../../../../common/framework/test_group.js';
 import { range, unreachable } from '../../../../../common/util/util.js';
 import { kMinDynamicBufferOffsetAlignment } from '../../../../capability_info.js';
+import { kResourceStates, ResourceState } from '../../../../gpu_test.js';
 import {
   kProgrammableEncoderTypes,
   ProgrammableEncoderType,
-} from '../../util/command_buffer_maker.js';
-import { kResourceStates, ResourceState, ValidationTest } from '../../validation_test.js';
+} from '../../../../util/command_buffer_maker.js';
+import { ValidationTest } from '../../validation_test.js';
 
 class F extends ValidationTest {
   encoderTypeToStageFlag(encoderType: ProgrammableEncoderType): GPUShaderStageFlags {
