+++ src/webgpu/api/validation/validation_test.ts
@@ -1,12 +1,5 @@
-import { assert, unreachable } from '../../../common/util/util.js';
 import { BindableResource, kMaxQueryCount } from '../../capability_info.js';
-import { GPUTest } from '../../gpu_test.js';
-
-import { CommandBufferMaker, EncoderType } from './util/command_buffer_maker.js';
-
-const kResourceStateValues = ['valid', 'invalid', 'destroyed'] as const;
-export type ResourceState = typeof kResourceStateValues[number];
-export const kResourceStates: readonly ResourceState[] = kResourceStateValues;
+import { GPUTest, ResourceState } from '../../gpu_test.js';
 
 /**
  * Base fixture for WebGPU validation tests.
@@ -288,161 +281,5 @@ export class ValidationTest extends GPUTest {
     });
     this.device.popErrorScope();
     return pipeline;
-  }
-
-  /**
-   * Returns a GPUCommandEncoder, GPUComputePassEncoder, GPURenderPassEncoder, or
-   * GPURenderBundleEncoder, and a `finish` method returning a GPUCommandBuffer.
-   * Allows testing methods which have the same signature across multiple encoder interfaces.
-   *
-   * @example
-   * ```
-   * g.test('popDebugGroup')
-   *   .params(u => u.combine('encoderType', kEncoderTypes))
-   *   .fn(t => {
-   *     const { encoder, finish } = t.createEncoder(t.params.encoderType);
-   *     encoder.popDebugGroup();
-   *   });
-   *
-   * g.test('writeTimestamp')
-   *   .params(u => u.combine('encoderType', ['non-pass', 'compute pass', 'render pass'] as const)
-   *   .fn(t => {
-   *     const { encoder, finish } = t.createEncoder(t.params.encoderType);
-   *     // Encoder type is inferred, so `writeTimestamp` can be used even though it doesn't exist
-   *     // on GPURenderBundleEncoder.
-   *     encoder.writeTimestamp(args);
-   *   });
-   * ```
-   */
-  createEncoder<T extends EncoderType>(
-    encoderType: T,
-    {
-      attachmentInfo,
-      occlusionQuerySet,
-    }: {
-      attachmentInfo?: GPURenderBundleEncoderDescriptor;
-      occlusionQuerySet?: GPUQuerySet;
-    } = {}
-  ): CommandBufferMaker<T> {
-    const fullAttachmentInfo = {
-      // Defaults if not overridden:
-      colorFormats: ['rgba8unorm'],
-      sampleCount: 1,
-      // Passed values take precedent.
-      ...attachmentInfo,
-    } as const;
-
-    switch (encoderType) {
-      case 'non-pass': {
-        const encoder = this.device.createCommandEncoder();
-
-        return new CommandBufferMaker(this, encoder, (shouldSucceed: boolean) =>
-          this.expectGPUError('validation', () => encoder.finish(), !shouldSucceed)
-        );
-      }
-      case 'render bundle': {
-        const device = this.device;
-        const rbEncoder = device.createRenderBundleEncoder(fullAttachmentInfo);
-        const pass = this.createEncoder('render pass', { attachmentInfo });
-
-        return new CommandBufferMaker(this, rbEncoder, (shouldSucceed: boolean) => {
-          // If !shouldSucceed, the resulting bundle should be invalid.
-          const rb = this.expectGPUError('validation', () => rbEncoder.finish(), !shouldSucceed);
-          pass.encoder.executeBundles([rb]);
-          // Then, the pass should also be invalid if the bundle was invalid.
-          return pass.validateFinish(shouldSucceed);
-        });
-      }
-      case 'compute pass': {
-        const commandEncoder = this.device.createCommandEncoder();
-        const encoder = commandEncoder.beginComputePass();
-
-        return new CommandBufferMaker(this, encoder, (shouldSucceed: boolean) => {
-          encoder.endPass();
-          return this.expectGPUError('validation', () => commandEncoder.finish(), !shouldSucceed);
-        });
-      }
-      case 'render pass': {
-        const makeAttachmentView = (format: GPUTextureFormat) =>
-          this.trackForCleanup(
-            this.device.createTexture({
-              size: [16, 16, 1],
-              format,
-              usage: GPUTextureUsage.RENDER_ATTACHMENT,
-              sampleCount: fullAttachmentInfo.sampleCount,
-            })
-          ).createView();
-
-        const passDesc: GPURenderPassDescriptor = {
-          colorAttachments: Array.from(fullAttachmentInfo.colorFormats, format => ({
-            view: makeAttachmentView(format),
-            loadValue: [0, 0, 0, 0],
-            storeOp: 'store',
-          })),
-          depthStencilAttachment:
-            fullAttachmentInfo.depthStencilFormat !== undefined
-              ? {
-                  view: makeAttachmentView(fullAttachmentInfo.depthStencilFormat),
-                  depthLoadValue: 0,
-                  depthStoreOp: 'discard',
-                  stencilLoadValue: 1,
-                  stencilStoreOp: 'discard',
-                }
-              : undefined,
-          occlusionQuerySet,
-        };
-
-        const commandEncoder = this.device.createCommandEncoder();
-        const encoder = commandEncoder.beginRenderPass(passDesc);
-        return new CommandBufferMaker(this, encoder, (shouldSucceed: boolean) => {
-          encoder.endPass();
-          return this.expectGPUError('validation', () => commandEncoder.finish(), !shouldSucceed);
-        });
-      }
-    }
-    unreachable();
-  }
-
-  /**
-   * Expect a validation error inside the callback.
-   *
-   * Tests should always do just one WebGPU call in the callback, to make sure that's what's tested.
-   */
-  expectValidationError(fn: () => void, shouldError: boolean = true): void {
-    // If no error is expected, we let the scope surrounding the test catch it.
-    if (shouldError) {
-      this.device.pushErrorScope('validation');
-    }
-
-    // Note: A return value is not allowed for the callback function. This is to avoid confusion
-    // about what the actual behavior would be; either of the following could be reasonable:
-    //   - Make expectValidationError async, and have it await on fn(). This causes an async split
-    //     between pushErrorScope and popErrorScope, so if the caller doesn't `await` on
-    //     expectValidationError (either accidentally or because it doesn't care to do so), then
-    //     other test code will be (nondeterministically) caught by the error scope.
-    //   - Make expectValidationError NOT await fn(), but just execute its first block (until the
-    //     first await) and return the return value (a Promise). This would be confusing because it
-    //     would look like the error scope includes the whole async function, but doesn't.
-    // If we do decide we need to return a value, we should use the latter semantic.
-    const returnValue = fn() as unknown;
-    assert(
-      returnValue === undefined,
-      'expectValidationError callback should not return a value (or be async)'
-    );
-
-    if (shouldError) {
-      const promise = this.device.popErrorScope();
-
-      this.eventualAsyncExpectation(async niceStack => {
-        const gpuValidationError = await promise;
-        if (!gpuValidationError) {
-          niceStack.message = 'Validation succeeded unexpectedly.';
-          this.rec.validationFailed(niceStack);
-        } else if (gpuValidationError instanceof GPUValidationError) {
-          niceStack.message = `Validation failed, as expected - ${gpuValidationError.message}`;
-          this.rec.debug(niceStack);
-        }
-      });
-    }
   }
 }
