+++ src/webgpu/util/command_buffer_maker.ts
@@ -1,4 +1,4 @@
-import { ResourceState, ValidationTest } from '../validation_test.js';
+import { ResourceState, GPUTest } from '../gpu_test.js';
 
 export const kRenderEncodeTypes = ['render pass', 'render bundle'] as const;
 export type RenderEncodeType = typeof kRenderEncodeTypes[number];
@@ -16,7 +16,7 @@ type EncoderByEncoderType<T extends EncoderType> = {
 }[T];
 
 /** See {@link webgpu/api/validation/validation_test.ValidationTest.createEncoder |
- * ValidationTest.createEncoder()}. */
+ * GPUTest.createEncoder()}. */
 export class CommandBufferMaker<T extends EncoderType> {
   /** `GPU___Encoder` for recording commands into. */
   // Look up the type of the encoder based on `T`. If `T` is a union, this will be too!
@@ -45,7 +45,7 @@ export class CommandBufferMaker<T extends EncoderType> {
   readonly validateFinishAndSubmitGivenState: (resourceState: ResourceState) => void;
 
   constructor(
-    t: ValidationTest,
+    t: GPUTest,
     encoder: EncoderByEncoderType<EncoderType>,
     finish: (shouldSucceed: boolean) => GPUCommandBuffer
   ) {
